Added some web monitor urls

+++ Test/ClearTaskTestFies.py
@@ -7,7 +7,8 @@ for log_file in pathlib.Path("Test/Log/").iterdir():
     os.remove(str(log_file))
 for task_json in pathlib.Path("Test/TestTasks/MainServerRoot/").iterdir():
     print("Removing: ", task_json)
-    os.remove(str(task_json))
+    if task_json.is_file():
+        os.remove(str(task_json))
 for task_dir in pathlib.Path("Test/TestTasks/ClientTaskRoot/").iterdir():
     print("Removing: ", task_dir)
     shutil.rmtree(str(task_dir))
